add balance

+++ src/App/Product/ProductFactory.java
@@ -4,7 +4,7 @@ package App.Product;
  * Created by orymamia on 19/12/2017.
  */
 public class ProductFactory {
-    public IProduct createProduct(String name, double price) {
+    public IProduct createProduct(String name, Integer price) {
         return new Product(name, price);
     }
 }
